Use raw_reaction_add for ReactionConfirmation
Previously, the `reaction_add` event was used, which required the members intent to be enabled in DMs for User objects to be updated. Therefore, `AdaptiveConfirmation` was made to switch between confirmation classes based on the bot's intents. However, I just realized the `raw_reaction_add` event provides just enough information DMs for reactions to work without needing members intent, so this event is now used. This also makes `AdaptiveConfirmation` obsolete, so it has been removed. It misused multiple inheritance anyways