Add "messagecount" command
Adds a message count command which shows the number of messages that have been sent in a server within the last 24 hours. This uses a `MessageTracker` cog which maintains its own in-memory database. On each message, it stores only the guild ID and the timestamp, and entries older than 24 hours are deleted every 12 hours.