Add lock to writing database connections
Adds `writing` parameter to `AsyncConnection` and `Database.connect()` that opts in to a queue for creating connections. This should be used for any connection that will be writing to the database to ensure that there is at most one connection active that intends to write. This prevents a possible error where one of multiple writing connections are blocked for too long since SQLite only supports one concurrent writer.