Fix "removenotes" command
The "removenotes" command makes multiple calls to `delete_note_by_note_id`, which assumes that the cache for a user will always exist and tries deleting it each time. After the first note is deleted though, the user is already removed from cache, so the second call would throw a KeyError.