Proper fix for "messagecount" graph
The actual issue was that matplotlib only distributed the bins across the bounds of the dataset, meaning if messages only occurred within the first hour, then instead of the 24x3 bins being spread across the 24 hours, it would use all those bins for that first hour instead. This fixes it by generating the bin edges manually. This also reverts last commit.