Add support for using reminders as announcements
Reminder content is no longer placed inside an embed, allowing both push notifications to display the content properly and for mentions to work. To limit abuse, mentions are escaped for users that do not have the Mention Everyone permission. See client_reminders_add() docstring for details on what features are offered for announcement creation.

More changes:
- Adds optional channel argument for "reminder add"
- Adds "announce" alias for "remind" command
- Increases max content for announcements to 1000
- Increases reminder cooldown by 1 second