Fix TypeError when backing in help command
When the help command is started on a cog or command and the user tries going back to BotPageSource, last_help is storing the bot mapping which matches the dict type, resulting in the view incorrectly treating the dict as kwargs.
This adds another condition to validate the dict keys' types.