Remove announcement-type reminders
This implementation was kind of hacky since it used the content of the message itself to distinguish announcement-type reminders. A preferable solution would be to have a specific announcement command and a boolean column in the database instead, that way they become independent of content and can have separate checks/cooldowns from regular reminders.